f prog adv fix

+++ functional_programming_advanced/filter.c
@@ -5,9 +5,9 @@
 size_t filter(int *array, size_t len, int **out_array, bool (*func)(int))
 {
     *out_array = malloc(sizeof(int *));
-    int out_i = 0;
+    size_t out_i = 0;
 
-    for (int i = 0; i < len; i++)
+    for (size_t i = 0; i < len; i++)
     {
         if (func(array[i]))
         {
